+++ app/src/main/java/stellarnear/aquene_dealer/Activities/MainActivityFragment.java
@@ -18,7 +18,7 @@ import android.widget.LinearLayout;
 
 import java.util.List;
 
-import stellarnear.aquene_dealer.Divers.QuadrantManager;
+import stellarnear.aquene_dealer.Divers.Quadrants.QuadrantManager;
 import stellarnear.aquene_dealer.R;
 
 /**
