+++ app/src/main/java/stellarnear/aquene_dealer/Divers/CombatPanel/CombatAskerResultLine.java
@@ -1,4 +1,4 @@
-package stellarnear.aquene_dealer.Divers;
+package stellarnear.aquene_dealer.Divers.CombatPanel;
 
 import android.app.Activity;
 import android.content.Context;
