+++ app/src/main/java/stellarnear/aquene_dealer/Divers/CombatPanel/CombatLauncherDamageLines.java
@@ -1,4 +1,4 @@
-package stellarnear.aquene_dealer.Divers;
+package stellarnear.aquene_dealer.Divers.CombatPanel;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -15,6 +15,11 @@ import android.widget.TextView;
 import java.util.Arrays;
 import java.util.List;
 
+import stellarnear.aquene_dealer.Divers.Rolls.ProbaFromDiceRand;
+import stellarnear.aquene_dealer.Divers.Rolls.DmgRoll;
+import stellarnear.aquene_dealer.Divers.Rolls.Roll;
+import stellarnear.aquene_dealer.Divers.Rolls.RollList;
+import stellarnear.aquene_dealer.Divers.Tools;
 import stellarnear.aquene_dealer.R;
 
 public class CombatLauncherDamageLines {
