+++ app/src/main/java/stellarnear/aquene_dealer/Divers/Rolls/DiceList.java
@@ -1,4 +1,4 @@
-package stellarnear.aquene_dealer.Divers;
+package stellarnear.aquene_dealer.Divers.Rolls;
 
 import java.util.ArrayList;
 import java.util.List;
