+++ app/src/main/java/stellarnear/aquene_dealer/Divers/Rolls/Roll.java
@@ -1,4 +1,4 @@
-package stellarnear.aquene_dealer.Divers;
+package stellarnear.aquene_dealer.Divers.Rolls;
 
 import android.content.Context;
 import android.widget.CheckBox;
